added my orderDetails function into utils and refactored accordingly
